Refactorizacion final, tests de integracion, servicios de TVDB...
